Troubleshooting update issue minor changes
urns out those that case difference between Email and email is what makes the difference. I pointed to the copy fl you made w/ the lower-cased the email field name and got the update to work fine going from SQL to AGOL. So we are all good there. Also, of note is that the name of the CSV is important to the feature layer of course, so you will want to note that in your documentation.I only made minor changes like adding an if to check for items in the updatefeatures before doing edit_features.